requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-telegram-bot
alembic
